lunatik: move object API to lunatik_obj.h
Extract object-related macros, inlines, and prototypes from lunatik.h
into a new lunatik_obj.h, included at the bottom of lunatik.h before
lunatik_val.h. Mirrors the lunatik_val.h pattern.
